adrv9002: Rename TX_ONLY to USE_RX_CLK_FOR_TX
The name USE_RX_CLK_FOR_TX has a correspondence in HDL and it better
describes the functionality. A Tx channel can be driven by a Rx clock
or it can be independent. It will make more sense when the Tx channel
driven by Rx clock will not be a matching channel number 1:1.
E.g. Rx1 clock could be used to drive Tx2 if the Rx2 interface is
disabled from HDL.

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>