adrv9002: Update profile check conditions
This update is correlated to a HDL update.
With this change, profiles that only enable Rx1-Tx2 and Rx2-Tx1 are
supported without being constrained to Txn clock/rate being equal with Rxn one.
Also, the HDL update adds the possibility for RX2 clock to be used for Tx1 and
Rx1 clock for Tx2, when the Rx SSI interface pair(Tx) is disabled.

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>