iio: adc: axi-adc: make use of devm_iio_dmaengine_buffer_setup()
For all the axi-adc drivers we have, there was no reason for passing
custom DMA ops when requesting the buffer. We were only cleaning some
status  bits that we really did not cared when buffering.

With the above we can remove code and use the (almost) identical API we
have upstream for setting up a DMA buffer.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>