iio: buffer-dma: sync with upstream
Now that upstream already supports high speed output buffers, the
missing piece is the new dma-buf interface. These change will sync as
much as possible with upstream code almost only leaving the legacy mmap
interface.

Note that the setup() api's have two extra arguments. AFAICS, they are
needed for the axi dac when the offload engine is to be used (or the
legacy GPIO based FIFO). So either we need to live with these minimal
changes or, ideally, we find an usecase to upstream the code together
with the offload engine (maybe it can be sold as a backend!)

Signed-off-by: Nuno Sa <nuno.sa@analog.com>